Avoid global references (#295)
Use a main function instead of allowing variables defined to be global
as this will make it easier in future to add tests using the main()
function as the entrypoint.

Switch to using altering the repository option to always end in a slash
and use the updated options dict where possible. Use explicit remote
variable in remaining instances.