Default to no repository defined
Don't assume a default repository name and require running scripts to
provide one should they wish to tag images with such a name and
subsequently use this as a target registry host.

Require repository to be provided so that it is non-empty until such
time as we also support pushing to docker hub and can check for any
arguments required in that case.

Change-Id: I84591cbf4ac4d0340d39d6ccd8752f1e83edfd0b