Harden the behaviour of the directory argument
So if I run tox -e windlass -- --directory $PWD
ns.directory will be '.' (tox does something) and
then os.path.split('.') returns the parent directory
which is not what I wanted.

Also if '/' is not added to the directory sometimes the
parent path is one up then what you expected.

Change-Id: Id0f185591fb851c4d2b368cfd3dd4f103aa71393