Remove unneeded features and ready windlass for new Python API
Removed  the --directory argument.

The --directory argument and semantics around finding the
products.yml files leads to confusion around where we are when
running Gather. This becomes a issue when we try and script
Gather as every script will have to either change directory
and sometimes it needs to be in the parent directory? In the
future we should use the repo: . relative to the yaml file
defining the artifact.

Remove :nowindlass docker tag semantics. This is way to specific
to docker container images then a more general solution to
managing artifacts.

Remove --repository argument. Gather already has the proxy-repository
which makes more sense to just use. Others depending on whether
push-only or build-only is set we may need up looking for different
container images in the docker registry.

When pulling or building an image we should just end up with an
image that matches the developer name in the yaml files. The alternative
is that we can have very inconsistent results.

If we want to push an image to a repositry use the proxy-repository
which will tag an temporarily from the developer image, push and
then remove this later.

Change-Id: If5d719ecd640798f65599e2e5d0bb2a4ce62594b