Add Gather Python API
The Python API allows use to treat a collection
of artifacts together and so we can script the management
of artifacts for a developer environment or for a CI
proposes.

We have a windlass.api.Gather object to manage a collection
of artifacts. This handles all management of running builds,
pushing images in parallel.

windlass.api.Artifact is the api to to manage an individual
artifact type.

windlass.products.Product is used to parse the input files
to read in the artifacts we wish to manage.

Change-Id: If7cfbf20daf204846e4847c94ee07fbcfc1d108e