Add option to download artifacts
This requires a overhaul to the windlass arguments. Had to
split out the repositories to download the artifacts from
to the repositories that we upload to.

--download-docker-image-registry and
--push-docker-image-registry

And the same for charts:

--download-charts-url and --push-charts-url

A version is required to download an artifact from.

Also added option --no-push to enforce that we do not
want to push artifacts to a repository.

Change-Id: Id056757dcc3cf54a9434e9d81d28852ef936acd9