Add fall_back mechanism to windlass
Allow failing back to different registries, so that
we can download from alpha, and fall back to downloading
from the staging repository for artifacts that haven't
been approved.

This change required to pull out the retry mechanism from
the windlass object and put it directly onto the artifacts.
This is so we can only retry one operation, as the windlass
cli would go and retry download / building and uploading
the artifact if it had a problem uploading

Also fixes bug when using no-parallel option. If we hit
an error during this we should fail early.

Change-Id: I21a81f4ff8ebdce6ea177dc5e1d898639a9a9873