Add mechanism to read / write artifact pins
Add api to read_pins and write_pins. These methods
take and return a list of Artifact objects. Each
artifact object will default to the version specified
in the manifest directory.

With the list of artifact objects we can windlass them
and promote at the end of a successful integration test.

Change-Id: I45b68f5bc764f5638af2686d6e7e2fb953fcdb6e