Artifact versioning
Make the management of artifacts versions more consistent
across artifacts.

Adds a version attribute to artifacts which can be
- None meaning that we use the version of the artifact
  specified in the repository or which is default for
  artifacts, like latest for images.
- Non-None meaning that we have specified a default
  version for the artifact. So if we upload the artifact
  we upload this version of the artifact.

When downloading or uploading an artifact we use the
version attribute unless we have override this via
the version argument to the appropriate methods.

What all this means is that we can download particular
versions of artifacts, and upload them under a different
version. We can also in the next change generate a set
of artifact objects from a manifest and we can
then download all the pinned artifacts, and upload them
to a different repository (promotion) and also in the
future this will probable be needed for integation
testing where we want to download all artifacts under
test and process them locally.

Add a --push-version argument to windlass and a
--download-version. This allows us to separate
the version we pull from to the version we upload.

Change-Id: I29f2d3d97a642fcc8e5780703787b7c5eb2ef5ea