Support devtag in image artifact names
This allows us to reference artifact like so
zing/percona:5.7 which makes integration with
docker-compose easier.

Note that we are going to have problems if we
tried to include two images with the same imagename
for example zing/percona:5.7 and zing/percona:5.8.
In this case we will build both images ok, but
publication is going to only publish one image and
it will not be clear which. This isn't really a
problem as we already have this problem.

Change-Id: I708a2ec89fd7ce9b8ed05f1772b8302ae5028d27