Improve upload of charts
During promotion we can end up trying to upload multiple
charts at the same time. This triggers Artifactory to regenerate
the index.yaml which is currently very racey. So avoid trying
to upload a chart if it already exists in Artifactory.

Add better logging to windlass. So we can see what charts were
published in the promotion jobs, and we can log errros when
things go wrong in debug mode.

Change-Id: I48f44e0d440aa4b1b7843dff2d4face3db66f5a9