We need the ability to ignore or include artifacts when pinning
When building the ncs/deployment image we have found charts
that need to be published to Zing but not pinned in the aws
directory but to the du directory, namely searchlight-du

This adds the ability to all artifact type to ignore certain
artifacts. It also does the inverse so we can specify 'only'
list of artifacts to include in a directory.

Change-Id: Ie870edc232691fdf9ac4314ad3360dfef9917cad