Have windlass multiprocessing worker return worker values
Returns a list with a return value for each artifact in the artifacts
list.  A None value is returned for any method which does not explicitly
return a value.

Tweaks windlass.py to not use the return value of Gather.run() since
this method wasn't returning a value before.

This is to allow the upload methods of Remote object to return the urls
they upload to.

Change-Id: I53b0836a35466dae42a62c32909fe8ccbcbaca1f