Align __repr__ and __str__ with Python guidelines
This will make __repr__ strings for artifacts to be able to be
parsed in python3's eval, so it would make debugging easier.
Other change is to make __str__ more human readable, which would
help with code reponsible for notifications, as it will no longer
have to deal with formatting strings.

Change-Id: Ibaee7ce8953a76764edada8b7eafbbd9a858ce46