Trottle number of parallel jobs in windlass
Use multiprocessing.pool.Pool to trottle the number of artifacts
we process. The size of this pool is controlable via the windlass
cli or the python api.

Error handling is done via the _er_cb callback which aborts any
other running builds

no-parallel is support by bypassing the pool altogether and just
calling the processor method directly.

Change-Id: Id9fc37e7a857d7b72c9493f6e40789eebca58535