The filename on a generic object can be explicit
When we donwload a generic artifact then we know the filename
of this artifact, and this filename is not the same as the filename
defined in the artifacts.yaml. In the artifacts.yaml this is a
regular experission for us to find the variable file.

This change will allow us to deal with generic artifacts during
promotion when we don't want to download the artifacts as this
can fill up the slaves disks

Change-Id: I68b4a668b43a7d954feb8c2b6eba661f8507e814