Add support for setting ECR lifecycle policy
Add support in the ECRConnector class for setting the lifecycle policy on
a new repository.
This is done by introducing a new 'repo_policies' initialisation parameter
in the position of the existing 'repo_policy' parameter, allowing the
new parameter to take either the old or new usage.
The 'repo_policy' parameter is kept to support any existing
named-parameter usage.

This should allow for deprecation of the 'repo_policy' parameter and its
eventual removal.

Includes some tests to ensure legacy usage is not broken.

Change-Id: Ifc7dfd3d80e704a2dcab59bca1cf3735a310807a