Make testing of container output more robust
Change from calling docker run with auto-removal to explicit container
creation in order to avoid a timing issue where the API might query for
the logs after the container was removed and not see the output.
Instead wait for the container to exit (using detach=False), and then
retrieve all log output, followed by removal of the container before
testing the returned output.