Upgrade docker python API to get ports support
The 'test.support' is an internal library of python which is not
guaranteed to be available in all distributions.

It is possible to use a lower level docker.APIClient to perform a
container inspection to retrieve dynamically allocated ports. As a more
recent version of the docker-py library has added support to retrieve
and store this information in an attributed, upgrade and switch.

See docker/docker-py#1742 for more details.