Support upload to multiple registries
Upload to multiple registries to allow publishing to multiple endpoints
where some may in turn do additional scanning of the images. Adds
support for retrieving separate authentication details from the
environment for each registry.

Use dedicated objects for docker registries to manage details on
connection and authentication. Migrate code to use this object and
requesting the connector which abstracts the authentication logic.

Due to using the DockerConnector and ECRConnector upload logic, replace
the remote_retry with the standard retry that is used on the images and
ensure that when a max retries exception is thrown that it simply
propagates up the stack instead of getting retried again.